+++ src/ex1/index.html
@@ -3,7 +3,8 @@
 
 <head>
     <title>Yuval Bader TDL</title>
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="./scripts/Styles/style.css" />
+    <link rel="stylesheet" href="./scripts/Styles/pokeball.css" />
 </head>
 
 <body>
@@ -12,27 +13,19 @@
         <div>
             <div class="add-new-container">
                 <input type="text" class="add-new-input" placeholder="Write your new Todo"> </input>
-                <!-- <button type="submit" class="add-new-button"> -->
                 <div class="pokeball add-new-button">
                     <div class="pokeball__button"></div>
                 </div>
-                <!-- </button> -->
             </div>
             <div class="todo-list-container">
                 <ul class="todo-list">
                 </ul>
-
             </div>
-
             <button class="delete-all-tasks">Delete all</button>
-
             <Text class="pending-tasks"> </Text>
-
         </div>
-
     </div>
     <script type="module" src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.esm.js"></script>
-    <script nomodule src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.js"></script>
 </body>
 <script type="module" src="scripts/main.js"></script>
