+++ src/ex1/scripts/Services/PokemonClient.js
