+++ src/ex1/scripts/Styles/style.css
@@ -9,7 +9,7 @@ body {
   background: whitesmoke;
   max-width: 600px;
   width: 100%;
-  height: 550px;
+  height: 570px;
   padding: 25px;
   border-radius: 5px;
   box-shadow: 0px 10px 15px rgba(0, 0, 0, 0.1);
@@ -18,7 +18,8 @@ body {
   position: relative;
   left: 50%;
   transform: translateX(-50%);
-  margin-top: 10%;
+  margin-top: 5%;
+  margin-bottom: 5%;
 }
 .todo-list-container {
   display: flex;
@@ -33,6 +34,7 @@ body {
     -1px 4px 0 #268d95, 1px 1px 0 #268d95;
   text-align: center;
 }
+
 .pending-tasks {
   font-size: large;
   margin-left: 19%;
@@ -123,7 +125,7 @@ body {
 }
 .add-new-container input {
   flex-grow: 1;
-  color: #bfd2ff;
+  color: #00b09b;
   font-size: 1.8rem;
   line-height: 2.4rem;
   vertical-align: middle;
@@ -165,95 +167,4 @@ body {
 .todo-list {
   overflow: auto;
   max-height: 260px;
-}
-.delete-all-tasks {
-}
-
-/* ///////////////// */
-
-/* Poké Styles */
-.pokeball {
-  position: relative;
-  width: 50px;
-  height: 50px;
-  background: #fff;
-  border: 2.5px solid #000;
-  border-radius: 50%;
-  overflow: hidden;
-  box-shadow: inset -2.5px 2.5px 0 2.5px #ccc;
-  animation: fall 0.25s ease-in-out,
-    shake 1.25s cubic-bezier(0.36, 0.07, 0.19, 0.97) 3;
-}
-.pokeball::before,
-.pokeball::after {
-  content: "";
-  position: absolute;
-}
-.pokeball::before {
-  background: red;
-  width: 100%;
-  height: 50%;
-}
-.pokeball::after {
-  top: calc(50% - 2.5px);
-  width: 100%;
-  height: 5px;
-  background: #000;
-}
-.pokeball__button {
-  position: absolute;
-  top: calc(50% - 7.5px);
-  left: calc(50% - 7.5px);
-  width: 15px;
-  height: 15px;
-  background: #7f8c8d;
-  border: 2.5px solid #fff;
-  border-radius: 50%;
-  z-index: 10;
-  box-shadow: 0 0 0 2.5px black;
-  animation: blink 0.5s alternate 7;
-}
-/* Animation */
-@keyframes blink {
-  from {
-    background: #eee;
-  }
-  to {
-    background: #e74c3c;
-  }
-}
-@keyframes shake {
-  0% {
-    transform: translate(0, 0) rotate(0);
-  }
-  20% {
-    transform: translate(-10px, 0) rotate(-20deg);
-  }
-  30% {
-    transform: translate(10px, 0) rotate(20deg);
-  }
-  50% {
-    transform: translate(-10px, 0) rotate(-10deg);
-  }
-  60% {
-    transform: translate(10px, 0) rotate(10deg);
-  }
-  100% {
-    transform: translate(0, 0) rotate(0);
-  }
-}
-
-@keyframes fall {
-  0% {
-    top: -200px;
-  }
-  60% {
-    top: 0;
-  }
-  80% {
-    top: -20px;
-  }
-  100% {
-    top: 0;
-  }
 }
